# lib/api-zod/src/generated/types/companySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry Farm Management API
 * OpenAPI spec version: 0.1.0
 */

export interface CompanySummary {
  totalFarms: number;
  totalHouses: number;
  activeCycles: number;
  completedCycles: number;
  totalChicksAllTime: number;
  totalRevenueAllTime: number;
  totalProfitAllTime: number;
  averageFcr: number;
  averageMortalityRate: number;
  bestFarmName?: string;
  worstFarmName?: string;
}