# lib/api-zod/src/generated/types/createCycleBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry Farm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCycleBodyBreed } from "./createCycleBodyBreed";
import type { CreateCycleBodyStatus } from "./createCycleBodyStatus";

export interface CreateCycleBody {
  cycleNumber: number;
  housingDate: Date;
  chickCount: number;
  chickPricePerUnit: number;
  breed: CreateCycleBodyBreed;
  totalFeedKg?: number;
  feedCostTotal?: number;
  totalMedicationCost?: number;
  totalMortality?: number;
  finalWeightKg?: number;
  saleAgedays?: number;
  salePricePerKg?: number;
  otherCosts?: number;
  status: CreateCycleBodyStatus;
}