# lib/api-zod/src/generated/types/cycle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Poultry Farm Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CycleBreed } from "./cycleBreed";
import type { CycleStatus } from "./cycleStatus";

export interface Cycle {
  id: number;
  houseId: number;
  cycleNumber: number;
  housingDate: Date;
  chickCount: number;
  chickPricePerUnit: number;
  breed: CycleBreed;
  totalFeedKg?: number;
  feedCostTotal?: number;
  totalMedicationCost?: number;
  totalMortality?: number;
  finalWeightKg?: number;
  saleAgedays?: number;
  salePricePerKg?: number;
  otherCosts?: number;
  status: CycleStatus;
  mortalityRate?: number;
  fcr?: number;
  averageWeightKg?: number;
  totalCost?: number;
  totalRevenue?: number;
  netProfit?: number;
  costPerLiveKg?: number;
  profitPerChick?: number;
}